// apps/client/lib/generated/worldItems.ts
// Generated by tools/rom-data/generate-world-content.py — do not edit by hand.
export const GENERATED_BAG_ITEMS = {
  "big-pearl": {
    "name": "Big Pearl",
    "firered": 107
  },
  "calcium": {
    "name": "Calcium",
    "firered": 67
  },
  "carbos": {
    "name": "Carbos",
    "firered": 66
  },
  "dire-hit": {
    "name": "Dire Hit",
    "firered": 74
  },
  "fire-stone": {
    "name": "Fire Stone",
    "firered": 95
  },
  "full-heal": {
    "name": "Full Heal",
    "firered": 23
  },
  "full-restore": {
    "name": "Full Restore",
    "firered": 19
  },
  "guard-spec": {
    "name": "Guard Spec.",
    "firered": 73
  },
  "hp-up": {
    "name": "Hp Up",
    "firered": 63
  },
  "iron": {
    "name": "Iron",
    "firered": 65
  },
  "leaf-stone": {
    "name": "Leaf Stone",
    "firered": 98
  },
  "max-elixir": {
    "name": "Max Elixir",
    "firered": 37
  },
  "max-potion": {
    "name": "Max Potion",
    "firered": 20
  },
  "max-repel": {
    "name": "Max Repel",
    "firered": 84
  },
  "max-revive": {
    "name": "Max Revive",
    "firered": 25
  },
  "nest-ball": {
    "name": "Nest Ball",
    "firered": 8
  },
  "net-ball": {
    "name": "Net Ball",
    "firered": 6
  },
  "nugget": {
    "name": "Nugget",
    "firered": 110
  },
  "pearl": {
    "name": "Pearl",
    "firered": 106
  },
  "poke-doll": {
    "name": "Poké Doll",
    "firered": 80
  },
  "pp-up": {
    "name": "Pp Up",
    "firered": 69
  },
  "protein": {
    "name": "Protein",
    "firered": 64
  },
  "super-repel": {
    "name": "Super Repel",
    "firered": 83
  },
  "thunder-stone": {
    "name": "Thunderstone",
    "firered": 96
  },
  "ultra-ball": {
    "name": "Ultra Ball",
    "firered": 2
  },
  "water-stone": {
    "name": "Water Stone",
    "firered": 97
  },
  "x-accuracy": {
    "name": "X Accuracy",
    "firered": 78
  },
  "x-defend": {
    "name": "X Defend",
    "firered": 76
  },
  "x-special": {
    "name": "X Special",
    "firered": 79
  },
  "x-speed": {
    "name": "X Speed",
    "firered": 77
  },
  "zinc": {
    "name": "Zinc",
    "firered": 70
  }
} as const;